@@ -12,42 +12,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.0.2",
+    "@angular/animations": "8.0.1",
     "@angular/cdk": "7.0.2",
-    "@angular/common": "7.0.2",
-    "@angular/compiler": "7.0.2",
-    "@angular/core": "7.0.2",
-    "@angular/forms": "7.0.2",
-    "@angular/http": "7.0.2",
+    "@angular/common": "8.0.1",
+    "@angular/compiler": "8.0.1",
+    "@angular/core": "8.0.1",
+    "@angular/forms": "8.0.1",
     "@angular/material": "7.0.2",
-    "@angular/platform-browser": "7.0.2",
-    "@angular/platform-browser-dynamic": "7.0.2",
-    "@angular/router": "7.0.2",
-    "@ngrx/effects": "^6.1.2",
-    "@ngrx/entity": "^6.1.2",
-    "@ngrx/store": "^6.1.2",
-    "@ngrx/store-devtools": "^6.1.2",
-    "angular-in-memory-web-api": "^0.7.0",
+    "@angular/platform-browser": "8.0.1",
+    "@angular/platform-browser-dynamic": "8.0.1",
+    "@angular/router": "8.0.1",
+    "angular-in-memory-web-api": "^0.8.0",
     "body-parser": "^1.18.2",
     "core-js": "^2.5.4",
     "dotenv": "^5.0.1",
     "express": "^4.16.3",
     "hammerjs": "^2.0.8",
     "mongoose": "^5.0.12",
-    "ngrx-data": "^6.1.0-beta.3",
     "rxjs": "^6.3.3",
     "subsink": "^0.1.0",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/cli": "^7.3.8",
-    "@angular/compiler-cli": "^7.2.14",
-    "@angular/language-service": "7.0.2",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "^8.0.3",
+    "@angular/compiler-cli": "^8.0.1",
+    "@angular/language-service": "8.0.1",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
@@ -58,6 +52,6 @@
     "protractor": "^5.4.1",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "3.1.6"
+    "typescript": "3.4.5"
   }
 }
