@@ -8,11 +8,11 @@ const routes: Routes = [
   { path: '', pathMatch: 'full', redirectTo: 'heroes' },
   {
     path: 'heroes',
-    loadChildren: './heroes/heroes.module#HeroesModule'
+    loadChildren: () => import('./heroes/heroes.module').then(m => m.HeroesModule)
   },
   {
     path: 'villains',
-    loadChildren: './villains/villains.module#VillainsModule',
+    loadChildren: () => import('./villains/villains.module').then(m => m.VillainsModule),
 
     // Solution - Apply the AuthGuard
     canActivate: [AuthGuard]
